+++ backend/src/common/enums/index.ts
@@ -1,5 +1,5 @@
 export * from './app';
 export * from './http';
 export * from './api';
-export * from './modelName';
 export * from './instance-type';
+export * from './model-name';
