--- backend/src/common/enums/modelName/index.ts
@@ -1,1 +1,0 @@
-export { ModelName } from 'healthcare-shared/common/enums';
