+++ shared/src/common/enums/index.ts
@@ -1,5 +1,4 @@
 export * from './http';
 export * from './exceptions';
 export * from './file';
-export * from './modelName';
 export * from './intance-type';
